requirements.txt
Flask==2.2.3
gunicorn==20.1.0
line-bot-sdk==2.4.2
python-dotenv==1.0.0
